third_party/WebKit/LayoutTests/fast/frames/form-submission-early-return-for-sandboxed-iframes.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<iframe id="iframe1" sandbox="allow-scripts allow-same-origin" src="resources/sandboxed-iframe-src.html"></iframe>
<script>
var validityChecked = false;
function setInvalid() { validityChecked = true; }
test (function() {
	document.getElementById('iframe1').onload = function() {
		window.frames['iframe1'].contentDocument.getElementById('submit1').click();
		assert_false(validityChecked);
	};
}, "Tests early return of sandboxed form without the need to check for validity.");
</script>
